pyproject.toml
[tool.poetry]
name = "pyai-caching"
version = "0.1.0"
description = "Redis-based caching for Pydantic AI LLM agents with cost tracking"
authors = ["Andrew de Jonge <talkingtoaj@hotmail.com>"]
readme = "README.md"
packages = [{include = "pyai_caching", from = "src"}]
license = "MIT"
keywords = ["llm", "caching", "redis", "ai", "cost-tracking"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = "^3.9"
redis = "^5.0.1"
pydantic = "^2.6.1"
typing-extensions = "^4.9.0"
pydantic-ai = "^0.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"
pytest-mock = "^3.12.0"
pytest-cov = "^4.1.0"
python-dotenv = "^1.0.1"
black = "^24.1.1"
ruff = "^0.2.1"
mypy = "^1.8.0"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py39"
select = ["E", "F", "I", "N", "W", "B", "UP", "PL", "RUF"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true 